requirements.txt
flask-restful
Flask-SQLAlchemy
SQLAlchemy
pysqlite
fabric
